requirements.txt
astroid==2.1.0
autopep8==1.4.3
beautifulsoup4==4.6.0
certifi==2019.3.9
chardet==3.0.4
cycler==0.10.0
idna==2.8
isort==4.3.4
kiwisolver==1.0.1
langdetect==1.0.7
langid==1.1.6
lazy-object-proxy==1.3.1
-e git+https://github.com/johnwmillr/LyricsGenius.git@ab46a32578ab74b4fa4125e6b2a14951ee485969#egg=lyricsgenius
matplotlib==3.0.2
mccabe==0.6.1
numexpr==2.6.9
numpy==1.15.4
pandas==0.23.4
Pillow==5.4.1
pip-autoremove==0.9.1
progressbar2==3.39.2
pycodestyle==2.4.0
pylint==2.2.2
pyparsing==2.3.0
python-dateutil==2.7.5
python-utils==2.3.0
pytz==2018.9
requests==2.21.0
scikit-learn==0.20.2
scipy==1.2.0
seaborn==0.9.0
six==1.12.0
tables==3.4.4
typed-ast==1.2.0
urllib3==1.24.1
wrapt==1.11.0
